factor node image provider code out and cleanup dead code
