+++ pkg/providers/imagefamily/image.go
@@ -17,36 +17,29 @@ limitations under the License.
 package imagefamily
 
 import (
-	"context"
-	"fmt"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v5"
-	"github.com/Azure/karpenter-provider-azure/pkg/apis/v1beta1"
 	"github.com/patrickmn/go-cache"
-	"github.com/samber/lo"
-	"k8s.io/client-go/kubernetes"
-	"sigs.k8s.io/karpenter/pkg/cloudprovider"
-	"sigs.k8s.io/karpenter/pkg/scheduling"
 	"sigs.k8s.io/karpenter/pkg/utils/pretty"
 )
 
-// TODO: Remove this provider, by refactoring it into the nodeimage.go provider, and resolver.go as needed
+// TODO: Remove this file, by refactoring remaining code into nodeimage.go (keeping for a follow up, as to not
+// combine code change, and moving bulk in the same PR, for cleaniness of reviewing)
+//
 // This is part of a shift to have the new controllers for k8s and node image version which populate status.
 // At the current point the provisioning logic isn't update itself, just the status being populated.
 // As part of the change to the actual provisioning, this provider will be refactored as mentioned above.
 // The logic the nodeimage.go provider is dependent upon will refactor into its file, and the runtime logic for
 // creation will refactor into resolver.go but dropping API retrievals for the data stored in the status instead.
 type Provider struct {
-	kubernetesVersionCache *cache.Cache
-	cm                     *pretty.ChangeMonitor
-	location               string
-	kubernetesInterface    kubernetes.Interface
-	imageCache             *cache.Cache
-	nodeImagesCache        *cache.Cache
-	imageVersionsClient    CommunityGalleryImageVersionsAPI
-	subscription           string
-	NodeImageVersions      NodeImageVersionsAPI
+	subscription string
+	location     string
+
+	imageVersionsClient CommunityGalleryImageVersionsAPI
+	nodeImageVersions   NodeImageVersionsAPI
+
+	nodeImagesCache *cache.Cache
+	cm              *pretty.ChangeMonitor
 }
 
 const (
@@ -57,71 +50,17 @@ const (
 	communityImageIDFormat          = "/CommunityGalleries/%s/images/%s/versions/%s"
 )
 
-func NewProvider(kubernetesInterface kubernetes.Interface, kubernetesVersionCache *cache.Cache, versionsClient CommunityGalleryImageVersionsAPI, location, subscription string, nodeImageVersionsClient NodeImageVersionsAPI) *Provider {
+func NewProvider(versionsClient CommunityGalleryImageVersionsAPI, location, subscription string, nodeImageVersionsClient NodeImageVersionsAPI) *Provider {
 	return &Provider{
-		kubernetesVersionCache: kubernetesVersionCache,
-		imageCache:             cache.New(imageExpirationInterval, imageCacheCleaningInterval),
-		nodeImagesCache:        cache.New(imageExpirationInterval, imageCacheCleaningInterval),
-		location:               location,
-		imageVersionsClient:    versionsClient,
-		cm:                     pretty.NewChangeMonitor(),
-		kubernetesInterface:    kubernetesInterface,
-		subscription:           subscription,
-		NodeImageVersions:      nodeImageVersionsClient,
+		subscription:        subscription,
+		location:            location,
+		imageVersionsClient: versionsClient,
+		nodeImageVersions:   nodeImageVersionsClient,
+		nodeImagesCache:     cache.New(imageExpirationInterval, imageCacheCleaningInterval),
+		cm:                  pretty.NewChangeMonitor(),
 	}
-}
-
-// TODO (charliedmcb): refactor this into resolver.go
-// resolveNodeImage returns Distro and Image ID for the given instance type. Images may vary due to architecture, accelerator, etc
-//
-// Preconditions:
-// - nodeImages is sorted by priority order
-func (r *defaultResolver) resolveNodeImage(nodeImages []v1beta1.NodeImage, instanceType *cloudprovider.InstanceType) (string, error) {
-	// nodeImages are sorted by priority order, so we can return the first one that matches
-	for _, availableImage := range nodeImages {
-		if err := instanceType.Requirements.Compatible(
-			scheduling.NewNodeSelectorRequirements(availableImage.Requirements...),
-			v1beta1.AllowUndefinedWellKnownAndRestrictedLabels,
-		); err == nil {
-			return availableImage.ID, nil
-		}
-	}
-	return "", fmt.Errorf("no compatible images found for instance type %s", instanceType.Name)
-}
-
-// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
-func (p *Provider) getCIGImageID(publicGalleryURL, communityImageName string) (string, error) {
-	imageVersion, err := p.latestNodeImageVersionCommunity(publicGalleryURL, communityImageName)
-	if err != nil {
-		return "", err
-	}
-	return BuildImageIDCIG(publicGalleryURL, communityImageName, imageVersion), nil
-}
-
-// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
-func (p *Provider) latestNodeImageVersionCommunity(publicGalleryURL, communityImageName string) (string, error) {
-	pager := p.imageVersionsClient.NewListPager(p.location, publicGalleryURL, communityImageName, nil)
-	topImageVersionCandidate := armcompute.CommunityGalleryImageVersion{}
-	for pager.More() {
-		page, err := pager.NextPage(context.Background())
-		if err != nil {
-			return "", err
-		}
-		for _, imageVersion := range page.CommunityGalleryImageVersionList.Value {
-			if lo.IsEmpty(topImageVersionCandidate) || imageVersion.Properties.PublishedDate.After(*topImageVersionCandidate.Properties.PublishedDate) {
-				topImageVersionCandidate = *imageVersion
-			}
-		}
-	}
-	return lo.FromPtr(topImageVersionCandidate.Name), nil
-}
-
-// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
-func BuildImageIDCIG(publicGalleryURL, communityImageName, imageVersion string) string {
-	return fmt.Sprintf(communityImageIDFormat, publicGalleryURL, communityImageName, imageVersion)
 }
 
 func (p *Provider) Reset() {
-	p.imageCache.Flush()
 	p.nodeImagesCache.Flush()
 }
