+++ pkg/providers/imagefamily/nodeimage.go
@@ -20,10 +20,12 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v5"
 	"github.com/Azure/karpenter-provider-azure/pkg/apis/v1beta1"
 	"github.com/Azure/karpenter-provider-azure/pkg/operator/options"
 	"github.com/mitchellh/hashstructure/v2"
 	"github.com/patrickmn/go-cache"
+	"github.com/samber/lo"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/karpenter/pkg/scheduling"
 )
@@ -80,7 +82,7 @@ func (p *Provider) List(ctx context.Context, nodeClass *v1beta1.AKSNodeClass) ([
 
 func (p *Provider) listSIG(ctx context.Context, supportedImages []DefaultImageOutput) ([]NodeImage, error) {
 	nodeImages := []NodeImage{}
-	retrievedLatestImages, err := p.NodeImageVersions.List(ctx, p.location, p.subscription)
+	retrievedLatestImages, err := p.nodeImageVersions.List(ctx, p.location, p.subscription)
 	if err != nil {
 		return nil, err
 	}
@@ -134,4 +136,36 @@ func (p *Provider) cacheKey(supportedImages []DefaultImageOutput, k8sVersion str
 		return "", err
 	}
 	return fmt.Sprintf("%016x", hash), nil
+}
+
+// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
+func (p *Provider) getCIGImageID(publicGalleryURL, communityImageName string) (string, error) {
+	imageVersion, err := p.latestNodeImageVersionCommunity(publicGalleryURL, communityImageName)
+	if err != nil {
+		return "", err
+	}
+	return BuildImageIDCIG(publicGalleryURL, communityImageName, imageVersion), nil
+}
+
+// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
+func (p *Provider) latestNodeImageVersionCommunity(publicGalleryURL, communityImageName string) (string, error) {
+	pager := p.imageVersionsClient.NewListPager(p.location, publicGalleryURL, communityImageName, nil)
+	topImageVersionCandidate := armcompute.CommunityGalleryImageVersion{}
+	for pager.More() {
+		page, err := pager.NextPage(context.Background())
+		if err != nil {
+			return "", err
+		}
+		for _, imageVersion := range page.CommunityGalleryImageVersionList.Value {
+			if lo.IsEmpty(topImageVersionCandidate) || imageVersion.Properties.PublishedDate.After(*topImageVersionCandidate.Properties.PublishedDate) {
+				topImageVersionCandidate = *imageVersion
+			}
+		}
+	}
+	return lo.FromPtr(topImageVersionCandidate.Name), nil
+}
+
+// TODO (charliedmcb): refactor this into nodeimage.go and create new provider
+func BuildImageIDCIG(publicGalleryURL, communityImageName, imageVersion string) string {
+	return fmt.Sprintf(communityImageIDFormat, publicGalleryURL, communityImageName, imageVersion)
 }
